fix (issue-239): updated storageId field type from int to long

+++ src/Crowdin.Api/SourceFiles/AddFileRequest.cs
@@ -11,7 +11,7 @@ namespace Crowdin.Api.SourceFiles
     public class AddFileRequest
     {
         [JsonProperty("storageId")]
-        public int StorageId { get; set; }
+        public long StorageId { get; set; }
         
         [JsonProperty("name")]
 #pragma warning disable 8618
